Rename: create utils and move formatDate and formatTimeDifference
