+++ src/components/CardList.js
@@ -1,35 +1,9 @@
 import "./CardList.css";
 import noImage from "../assets/noImage.svg";
+import formatDate from "../utils/formatDate.js";
+import formatTimeDifference from "../utils/formatTimeDifference.js";
 
-function formatDate(value) {
-  const date = new Date(value);
-  return `${date.getFullYear()}. ${date.getMonth() + 1}. ${date.getDate()}`;
-}
-
-function formatTimeDifference(value) {
-  const now = new Date();
-  const diffInMilliseconds = now - new Date(value);
-
-  const units = [
-    { label: "year", duration: 365 * 24 * 60 * 60 * 1000 },
-    { label: "month", duration: 30 * 24 * 60 * 60 * 1000 },
-    { label: "day", duration: 24 * 60 * 60 * 1000 },
-    { label: "hour", duration: 60 * 60 * 1000 },
-    { label: "minute", duration: 60 * 1000 },
-  ];
-
-  for (const unit of units) {
-    const diffInUnit = Math.floor(diffInMilliseconds / unit.duration);
-
-    if (diffInUnit === 1) {
-      return `1 ${unit.label} ago`;
-    } else if (diffInUnit > 1) {
-      return `${diffInUnit} ${unit.label}s ago`;
-    }
-  }
-}
-
-function FolderListItem({ item }) {
+function CardListItem({ item }) {
   const { title, createdAt, url, description, imageSource } = item;
 
   return (
@@ -49,13 +23,13 @@ function FolderListItem({ item }) {
   );
 }
 
-function FolderList({ items }) {
+function CardList({ items }) {
   return (
     <div className="card-container-center">
       <div className="card-container">
         {items.map((item) => (
           <div key={item.id} className="card-box">
-            <FolderListItem item={item} />
+            <CardListItem item={item} />
           </div>
         ))}
       </div>
@@ -63,4 +37,4 @@ function FolderList({ items }) {
   );
 }
 
-export default FolderList;
+export default CardList;
